pyproject.toml
[project]
name = "faust-macrogen"
version = "0.9.2"
description = "Macrogenetic Analysis for the Faust edition"
authors = [{ name = "Thorsten Vitt", email = "thorsten.vitt@uni-wuerzburg.de" }]
license = { text = "GPL-3.0" }
readme = "README.md"
requires-python = "<4.0,>=3.9"
dependencies = [
  "numpy<2.0.0,>=1.22.4",
  "more-itertools<9.0.0,>=8.0.2",
  "networkx<2.6,>=2.5",
  "pygraphviz<2.0,>=1.5",
  "ruamel-yaml<1.0.0,>=0.16.5",
  "pyyaml>=6.0,<7.0",
  "pandas<2.0.0,>=1.2.2",
  "openpyxl<4.0.0,>=3.0.2",
  "xlrd<3.0.0,>=2.0.1",
  "lxml<5.0.0,>=4.9.1",
  "requests<3.0.0,>=2.22.0",
  "requests-cache<1.0.0,>=0.5.2",
  "logging-tree<2.0,>=1.8",
  "colorlog<5.0.0,>=4.1.0",
  "tqdm<5.0.0,>=4.41.1",
  "pydantic-settings>=2.6.1",
]

[project.scripts]
witness-report = "macrogen.uris:_witness_report"
macrogen-diff = "macrogen.diff:main"
macrogen = "macrogen.main:main"
macrogen-viewer-dev = "graphviewer.gvfa:run_dev_server"

[project.optional-dependencies]
igraph = ["python-igraph<1.0.0,>=0.8.3"]
oldgraphviewer = ["flask<2.0.0,>=1.1.1"]
production = [
  "fastapi[all]<1.0.0,>=0.78.0",
  "gunicorn<21.0.0,>=20.0.4",
  "Jinja2<4.0.0,>=3.1.2",
]
solver = ["cvxpy<2.0.0,>=1.0.25", "cvxopt<2.0.0,>=1.2.3"]

[dependency-groups]
dev = ["pytest>=6.2.2", "coverage>=5.0.3", "watchdog>=0.10.3"]

[tool.hatch.build]
skip-excluded-dirs = true

[tool.hatch.build.targets.wheel]
packages = ["src/macrogen", "src/graphviewer"]

[tool.hatch.build.targets.sdist]
exclude = ["/data"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
